+++ template/build-scripts/compile.cjs.after.mjs
@@ -1,0 +1,4 @@
+(async () => {
+    console.info("[CJS compile post-processing started]");
+    console.info("[CJS compile post-processing ended]");
+})();