+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 (nothing yet)
 
+## [3.0.1] - 2021-02-26
+### Changed
+- updated eslint config
+### Fixed
+- build script failing
+
 ## [3.0.0] - 2021-02-19
 ### Added
 - migrations support